refactor : Refactor register
