+++ src/libs/utils/constants.ts
@@ -1,3 +1,4 @@
 export const SERVER_URL = 'https://affluent-closet.herokuapp.com';
+
 export const MAIN_URL =
   process.env.REACT_APP_MAIN_URL || 'http://localhost:3000';
