+++ src/model/enums.ts
@@ -8,3 +8,9 @@ export enum Gender {
   MAIL = 'MAIL',
   FEMAIL = 'FEMAIL',
 }
+
+export enum LOCAL_STORAGE_KEY {
+  USER_TOKEN = 'USER',
+  USER_BASE_INFO = 'userBaseInfo',
+  USER_REMIND_INFO = 'userRemindInfo',
+}
